+++ escomsoleje/EscomsoleJE.java
@@ -17,8 +17,8 @@ public class EscomsoleJE {
             if (status == Interpreter.FILE_NOT_FOUND)
                 System.err.println(String.format("File '%s' not found\n", new File(args[0]).getAbsolutePath()));
             
-            if (status == 0)
-                System.out.println("Valid program");
+//            if (status == 0)
+//                System.out.println("Valid program");
             
             System.exit(status);
         }